+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 
+* `build_relationship` function to simplify creating relationships in parallel
+
 ## 0.4.0
 
 ### Changed